OP#9528: Extract `TrackParser` and related classes into a new `parser.track_parser` module and update imports
