+++ tests/unit/OTAnalytics/application/test_datastore.py
@@ -7,14 +7,10 @@ import pytest
 from numpy import array, int32
 from PIL import Image
 
-from OTAnalytics.application.datastore import (
-    Datastore,
-    EventListParser,
-    TrackParser,
-    VideoParser,
-)
+from OTAnalytics.application.datastore import Datastore, EventListParser, VideoParser
 from OTAnalytics.application.parser.config_parser import ConfigParser
 from OTAnalytics.application.parser.flow_parser import FlowParser
+from OTAnalytics.application.parser.track_parser import TrackParser
 from OTAnalytics.domain.event import EventRepository
 from OTAnalytics.domain.flow import FlowRepository
 from OTAnalytics.domain.progress import ProgressbarBuilder
