+++ tests/unit/OTAnalytics/plugin_parser/test_convert_ottrk_to_feathers.py
@@ -3,7 +3,10 @@
 from unittest.mock import Mock
 
 import OTAnalytics.plugin_parser.ottrk_dataformat as ottrk_format
-from OTAnalytics.application.datastore import DetectionMetadata, TrackParseResult
+from OTAnalytics.application.parser.track_parser import (
+    DetectionMetadata,
+    TrackParseResult,
+)
 from OTAnalytics.domain.georeference import GeoreferenceMetadata
 from OTAnalytics.domain.video import VideoMetadata
 from OTAnalytics.plugin_parser.convert_ottrk_to_feathers import create_metadata_dict
